Add support for custom table filters
